Create 3. Longest Substring Without Repeating Characters
Two pointer or slide window, miantain the set with every extinct character existing only one time. Moving forward the right pointer, if meet character which is not in set , add it in the set, else if the character is already in the set, moving the left pointer and delete the corresponding character in the set. Repeate this process  until right pointer character not in the set, and add it in the set and moving forward. Every repeatition will maintain extinct character show up only once in the window. So in every loop, it should count the length of the set and update  the max length of substring